docker: Specify amd64 platform
This tells Docker to use the x86-64 port of Ubuntu in a full VM rather
than trying to run the native port of Ubuntu.

We need this because we are leaning on the host GCC to build x86 code
rather than doing a full cross compiler.

Thank you Victor Zimmer for the suggestion.